requirements.txt
numpy==1.19.5
Pillow==7.1.2
responder==2.0.5